requirements.txt
pandas
numpy
pydantic
nltk
tqdm
markitdown[all]
torch
transformers
google-genai
together
pydub
accelerate 
torch
peft